% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_sim.R
\name{compensator.hawkes}
\alias{compensator.hawkes}
\title{Compensator of a simulated Hawkes process}
\usage{
\method{compensator}{hawkes}(hawkes, t, ...)
}
\arguments{
\item{hawkes}{A simulated Hawkes process.}

\item{t}{A numeric or vector.}
}
\value{
The compensator at time t.
}
\description{
Outputs the compensator (integrated intensity) of a simulated Hawkes process.
}
\examples{
# Simulate an exponential Hawkes process with baseline intensity 1 and 
# excitation function 1*exp(-2t)
x <- hawkes(10, 1, 1, 2)
plot(x)
compensator(x, 0:10)
}
