% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inc_gamma_imag}
\alias{inc_gamma_imag}
\title{Incomplete gamma function of imaginary argument}
\usage{
inc_gamma_imag(nu, r)
}
\arguments{
\item{nu}{A number between 0 and 1 (strictly)}

\item{r}{A non-negative number}
}
\value{
The incomplete gamma function of imaginary argument (see Details)
}
\description{
Calculates the value of
\deqn{\Gamma_1(\nu, r) = \int_r^\infty y^{\nu-1} e^{-iy} \mathrm{d}y}
for \eqn{0 < \nu < 1} through the following relation (obtained by contour integration)
\deqn{\int_r^\infty y^{\nu-1} e^{-iy} \mathrm{d}y =
e^{-i\frac{\pi}{2}\nu} \int_r^\infty x^{\nu-1} e^{-x} \mathrm{d}x -
e^{-i\frac{\pi}{2}(\nu-1)} r^\nu \int_0^{\pi/2}
e^{i\theta \nu}e^{-re^{i\theta}}\mathrm{d}\theta.}
The first integral is calculated using function "tgamma" from the library
"boost::math", while the second is approximated via Simpson's rule.
}
\examples{
inc_gamma_imag(0.5, 1.0)
}
