% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_sim.R
\name{intensity.hawkes}
\alias{intensity.hawkes}
\title{Intensity of a simulated Hawkes process}
\usage{
\method{intensity}{hawkes}(hawkes, t)
}
\arguments{
\item{hawkes}{A simulated Hawkes process.}

\item{t}{A numeric or vector.}
}
\value{
The intensity at time t.
}
\description{
Outputs the intensity of a simulated Hawkes process.
}
\examples{
# Simulate an exponential Hawkes process with baseline intensity 1 and 
# excitation function 1*exp(-2t)
x <- hawkes(10, 1, 1, 2)
plot(x)
intensity(x, 0:10)
}
