% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_ogata.R
\name{intensity_ogata}
\alias{intensity_ogata}
\title{Intensity of a simulated Hawkes process}
\usage{
intensity_ogata(hawkes, t)
}
\arguments{
\item{hawkes}{A simulated Hawkes process from \code{hawkes_ogata}.}

\item{t}{A numeric or vector.}
}
\value{
The intensity at time t.
}
\description{
Outputs the intensity of a simulated Hawkes process.
}
\examples{
# Simulate an exponential Hawkes process with baseline intensity 1 and
# excitation function 1*exp(-2t)
x <- hawkes_ogata(10, 1, 1, 2)
plot(x)
intensity_ogata(x, 0:10)
}
