% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_mult.R
\name{matrix_mult}
\alias{matrix_mult}
\title{Simple matrix multiplication}
\usage{
matrix_mult(A, B)
}
\arguments{
\item{A}{Matrix}

\item{B}{Matrix}
}
\value{
Product of matrices
}
\description{
Simple matrix multiplication
}
\examples{
A <- matrix(1:9, 3, 3)
B <- matrix(11:19, 3, 3)
matrix_mult(A, B)
}
