% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Fitting Hawkes processes from continuous data}
\usage{
mle(model, events, end, opts = NULL, ...)
}
\arguments{
\item{model}{An object of class Model}

\item{events}{The locations of events (sorted in ascending order)}

\item{end}{The time until which the process is observed.}

\item{opts}{(Optional) To be passed to \code{nloptr}}

\item{...}{Additional arguments passed to \code{nloptr}}
}
\value{
Returns the \code{optim} output
}
\description{
This function fits a Hawkes process to continuous data by minimizing the likelihood
on the interval \eqn{[0,T]}.
}
\examples{
x = hawkes(100, fun = 1, repr = .5, family = "exp", rate = 1)
model = new(Exponential)
mle(model, x$p, x$T)
}
