% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_sim.R
\name{plot.hawkes}
\alias{plot.hawkes}
\title{Plot of a simulated Hawkes process}
\usage{
\method{plot}{hawkes}(hawkes, precision = 1000, ...)
}
\arguments{
\item{hawkes}{A simulated Hawkes process.}

\item{precision}{(default = 1e3) Number of points to plot.}
}
\value{
None
}
\description{
Plots a simulated Hawkes process, highlighting the steps used in Ogata's thinning algorithm.
}
\examples{
# Simulate an exponential Hawkes process with baseline intensity 1 and 
# excitation function 1*exp(-2t)
x <- hawkes(10, 1, 1, 2)
plot(x)
}
