% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_sim.R
\name{residuals.hawkes}
\alias{residuals.hawkes}
\title{Residuals of a simulated Hawkes process}
\usage{
\method{residuals}{hawkes}(hawkes)
}
\arguments{
\item{hawkes}{A simulated Hawkes process.}
}
\value{
The intensity at time t.
}
\description{
Outputs the residuals (values of the compensator at the times of arrival) of a simulated Hawkes process.
Useful function for diagnosis through the random time change theorem: the residuals should follow
a unit rate Poisson process.
}
\examples{
# Simulate an exponential Hawkes process with baseline intensity 1 and 
# excitation function 1*exp(-2t)
x <- hawkes(10, 1, 1, 2)
plot(x)
z = residuals(x)
plot(z)
abline(0, 1)
}
