% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_ogata.R
\name{residuals.hawkes_ogata}
\alias{residuals.hawkes_ogata}
\title{Residuals of a simulated Hawkes process}
\usage{
residuals.hawkes_ogata(hawkes)
}
\arguments{
\item{hawkes}{A simulated Hawkes process from \code{hawkes_ogata}.}
}
\value{
The intensity at time t.
}
\description{
Outputs the residuals (values of the compensator at the times of arrival) of a simulated Hawkes process.
Useful function for diagnosis through the random time change theorem: the residuals should follow
a unit rate Poisson process.
}
\examples{
# Simulate an exponential Hawkes process with baseline intensity 1 and
# excitation function 1*exp(-2t)
x <- hawkes_ogata(10, 1, 1, 2)
plot(x)
z = residuals(x)
plot(z)
abline(0, 1)
}
