% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Fitting Hawkes processes from continuous data}
\usage{
mle(model, data, opts = NULL, ...)
}
\arguments{
\item{model}{An object of class Model}

\item{data}{An object of class ContinuousData}

\item{opts}{(Optional) To be passed to \code{nloptr}}

\item{...}{Additional arguments passed to \code{nloptr}}
}
\value{
Returns the \code{optim} output
}
\description{
This function fits a Hawkes process to continuous data by minimizing the likelihood.
}
\examples{
x = hawkes(100, fun = 1, repr = .5, family = "exp", rate = 1)
model = new(Exponential)
data = new(ContinuousData, x$p, x$T)
mle(model, data)
}
