% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittle.R
\name{whittle}
\alias{whittle}
\title{Fitting Hawkes processes from discrete data}
\usage{
whittle(model, counts, binsize, trunc = 5L, ...)
}
\arguments{
\item{model}{An object of class Model}

\item{counts}{A bin-count sequence}

\item{binsize}{The bin size of the bin-count sequence}

\item{trunc}{(Optional) The number of foldings taken into account due to aliasing}

\item{...}{Additional arguments passed to \code{optim}}
}
\value{
Returns the \code{optim} output
}
\description{
This function fits a Hawkes process to discrete data by minimizing the Whittle contrast.
}
\examples{
x = hawkes(100, fun = 1, repr = .5, family = "exp", rate = 1)
y = discrete(x, binsize = 1)
model = new(Exponential)
whittle(model, y, 1)
}
