% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittle.R
\name{whittle}
\alias{whittle}
\title{Fitting Hawkes processes from discrete data}
\usage{
whittle(counts, kern, binsize = 1, trunc = 5L, init = NULL, ...)
}
\arguments{
\item{counts}{A bin-count sequence}

\item{kern}{Either a string (partially) matching one of the kernels implemented (see Details), or an object of class Model}

\item{binsize}{(Optional) The bin size of the bin-count sequence; ignored if \code{kern} is of class Model}

\item{trunc}{(Optional) The number of foldings taken into account due to aliasing}

\item{init}{(Optional) Initial values of the optimisation algorithm}

\item{...}{Additional arguments passed to \code{optim}}
}
\value{
eturns a list containing the solution of the optimisation procedure, the object Model
with its parameters updated to the solution, and the output produced by \code{optim}.
}
\description{
This function fits a Hawkes process to discrete data by minimizing the Whittle contrast.
}
\details{
If specified as string, the argument \code{kern} must match (partially) one of the following
(upper cases not taken into account): Exponential, SymmetricExponential,
Gaussian, PowerLaw, Pareto3, Pareto2, Pareto1.
The periodogram used in the optimisation procedure is computed in complexity
\eqn{O(n \log n)}, using function \code{fft}.
}
\examples{
# Simulate and fit a Hawkes process with exponential kernel
x = hawkes(100, fun = 1, repr = .5, family = "exp", rate = 1)
y = discrete(x, binsize = 1)
whittle(y, "Exponential")

# Simulate and fit a Hawkes process with power law kernel
x = hawkes(100, fun = 1, repr= .3, family = "powerlaw", shape = 3.5, scale = 1.0)
y = discrete(x, binsize = 1)
whittle(y, "powerlaw", 1.0, trunc = 2L, init = c(2.0, .5, 3.0, 1.5))
}
